refactor: backend update and delete products challenge
