--- sessions/backend-update-and-delete/products/components/ProductList/ProductList.styled.js
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-export const StyledHeading = styled.h2`
-  text-align: center;
-  color: var(--color-nemo);
-`;
-
-export const StyledList = styled.ul`
-  list-style-type: none;
-  display: grid;
-  gap: 1rem;
-  justify-items: center;
-  padding: 0;
-`;
