+++ sessions/backend-update-and-delete/products/pages/api/products/[id].js
@@ -9,9 +9,13 @@ export default async function handler(request, response) {
     const product = await Product.findById(id).populate("reviews");
 
     if (!product) {
-      return response.status(404).json({ status: "Not Found" });
+      response.status(404).json({ status: "Not Found" });
+      return;
     }
 
     response.status(200).json(product);
+    return;
   }
+
+  response.status(405).json({ status: "Method not allowed." });
 }
