+++ sessions/backend-update-and-delete/products_solution/components/Button.js
@@ -1,20 +1,13 @@
 import styled from "styled-components";
-import Link from "next/link";
-export const StyledLink = styled(Link)`
-  text-decoration: none;
-  display: inline-block;
-  color: inherit;
+
+export default styled.button`
+  appearance: none;
   border: none;
   background: var(--color-water-10);
   font-size: larger;
   padding: 0.5rem 1rem;
   box-shadow: 0px 1px 5px -2px var(--color-granite);
-
   &:hover {
     cursor: pointer;
-  }
-
-  &:visited {
-    color: inherit;
   }
 `;
