--- sessions/backend-update-and-delete/products_solution/components/Product/index.js
@@ -1,78 +1,0 @@
-import useSWR from "swr";
-import { useRouter } from "next/router";
-import { ProductCard } from "./Product.styled";
-import { StyledLink } from "../Link/Link.styled";
-import Comments from "../Comments";
-import ProductForm from "../ProductForm";
-import { StyledButton } from "../Button/Button.styled";
-import { useState } from "react";
-
-export default function Product() {
-  const [isEditMode, setIsEditMode] = useState(false);
-  const router = useRouter();
-  const { id } = router.query;
-
-  const { data, isLoading, mutate } = useSWR(`/api/products/${id}`);
-
-  async function handleEditProduct(productData) {
-    const response = await fetch(`api/products/${id}`, {
-      method: "PUT",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(productData),
-    });
-
-    if (!response.ok) {
-      console.error(response.status);
-      return;
-    }
-
-    mutate();
-    setIsEditMode(false);
-  }
-
-  async function handleDeleteProduct() {
-    const response = await fetch(`api/products/${id}`, {
-      method: "DELETE",
-    });
-
-    if (!response.ok) {
-      console.error(response.status);
-      return;
-    }
-
-    router.push("/");
-  }
-
-  if (isLoading) {
-    return <h1>Loading...</h1>;
-  }
-
-  if (!data) {
-    return;
-  }
-
-  return (
-    <ProductCard>
-      <h2>{data.name}</h2>
-      <p>Description: {data.description}</p>
-      <p>
-        Price: {data.price} {data.currency}
-      </p>
-      {data.reviews.length > 0 && <Comments reviews={data.reviews} />}
-      <StyledButton onClick={() => setIsEditMode(!isEditMode)}>
-        edit
-      </StyledButton>
-      <StyledButton onClick={handleDeleteProduct}>Delete</StyledButton>
-      {isEditMode && (
-        <ProductForm
-          onSubmit={handleEditProduct}
-          values={data}
-          isEditMode={true}
-        />
-      )}
-      <StyledLink href="/">Back to all</StyledLink>
-    </ProductCard>
-  );
-}
