+++ sessions/backend-update-and-delete/products_solution/components/ProductForm.js
@@ -1,5 +1,5 @@
-import { StyledForm, StyledHeading, StyledLabel } from "./ProductForm.styled";
-import { StyledButton } from "../Button/Button.styled";
+import styled from "styled-components";
+import StyledButton from "@/components/Button";
 
 const DEFAULT_VALUES = {
   name: "",
@@ -64,3 +64,20 @@ export default function ProductForm({
     </StyledForm>
   );
 }
+
+export const StyledForm = styled.form`
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+`;
+
+export const StyledHeading = styled.h2`
+  text-align: center;
+  color: var(--color-nemo);
+`;
+
+export const StyledLabel = styled.label`
+  display: flex;
+  justify-content: space-between;
+  gap: 5px;
+`;
