+++ sessions/backend-update-and-delete/products_solution/pages/[id].js
@@ -1,9 +1,83 @@
-import Product from "../components/Product";
+import { useState } from "react";
+import useSWR from "swr";
+import { useRouter } from "next/router";
+import styled from "styled-components";
+import StyledLink from "@/components/Link/";
+import Comments from "@/components/Comments";
+import ProductForm from "@/components/ProductForm";
+import StyledButton from "@/components/Button";
 
-export default function ProductDetailsPage() {
+export default function Product() {
+  const [showEditForm, setShowEditForm] = useState(false);
+  const router = useRouter();
+  const { id } = router.query;
+
+  const { data, isLoading, mutate } = useSWR(`/api/products/${id}`);
+
+  async function handleEditProduct(productData) {
+    const response = await fetch(`api/products/${id}`, {
+      method: "PUT",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(productData),
+    });
+
+    if (!response.ok) {
+      console.error(response.status);
+      return;
+    }
+
+    mutate();
+    setShowEditForm(false);
+  }
+
+  async function handleDeleteProduct() {
+    const response = await fetch(`api/products/${id}`, {
+      method: "DELETE",
+    });
+
+    if (!response.ok) {
+      console.error(response.status);
+      return;
+    }
+
+    router.push("/");
+  }
+
+  if (isLoading) {
+    return <h1>Loading...</h1>;
+  }
+
+  if (!data) {
+    return;
+  }
+
   return (
-    <>
-      <Product />
-    </>
+    <ProductCard>
+      <h2>{data.name}</h2>
+      <p>Description: {data.description}</p>
+      <p>
+        Price: {data.price} {data.currency}
+      </p>
+      {data.reviews.length > 0 && <Comments reviews={data.reviews} />}
+      <StyledButton onClick={() => setShowEditForm(!showEditForm)}>
+        edit
+      </StyledButton>
+      <StyledButton onClick={handleDeleteProduct}>Delete</StyledButton>
+      {showEditForm && (
+        <ProductForm
+          onSubmit={handleEditProduct}
+          values={data}
+          isEditMode={true}
+        />
+      )}
+      <StyledLink href="/">Back to all</StyledLink>
+    </ProductCard>
   );
 }
+
+export const ProductCard = styled.article`
+  padding: 0.5rem 1rem;
+  box-shadow: 0px 1px 5px -2px var(--color-granite);
+`;
