+++ sessions/backend-update-and-delete/products_solution/pages/index.js
@@ -1,12 +1,7 @@
-import ProductList from "../components/ProductList";
+import ProductList from "@/components/ProductList";
 import styled from "styled-components";
-import ProductForm from "../components/ProductForm";
+import ProductForm from "@/components/ProductForm";
 import useSWR from "swr";
-
-const Heading = styled.h1`
-  text-align: center;
-  color: var(--color-nemo);
-`;
 
 export default function HomePage() {
   const { mutate } = useSWR("/api/products");
@@ -41,3 +36,8 @@ export default function HomePage() {
     </>
   );
 }
+
+const Heading = styled.h1`
+  text-align: center;
+  color: var(--color-nemo);
+`;
